Fix signed publishing and maven central validation failures.
Fix publishing for GPG2 by going back to v1.1.0 of `sbt-pgp` (see
comments on PR at https://github.com/sbt/sbt-pgp/pull/117). We'll go
forward again when a fix is present.

The SCM is now added automatically to the POM by newer versions of SBT,
so this causes a validation failure when publishing to maven central.